+++ src/cpprealm/internal/bridge/sync_session.cpp
@@ -1,9 +1,12 @@
 #include <cpprealm/internal/bridge/sync_session.hpp>
+#include <cpprealm/internal/bridge/status.hpp>
+
 #include <realm/object-store/sync/sync_session.hpp>
 
 #include <future>
 
 namespace realm::internal::bridge {
+
     enum sync_session::state sync_session::state() const {
         if (auto session = m_session.lock()) {
             return static_cast<enum sync_session::state>(session->state());
@@ -11,20 +14,20 @@ namespace realm::internal::bridge {
         throw std::runtime_error("Realm: Error accessing sync_session which has been destroyed.");
     }
 
-    void sync_session::wait_for_download_completion(std::function<void(std::error_code)> &&callback) {
+    void sync_session::wait_for_download_completion(std::function<void(status)> &&callback) {
         if (auto session = m_session.lock()) {
-            session->wait_for_download_completion([cb = std::move(callback)](auto ec) {
-                cb(ec.get_std_error_code());
+            session->wait_for_download_completion([cb = std::move(callback)](::realm::Status s) {
+                cb(std::move(s));
             });
         } else {
             throw std::runtime_error("Realm: Error accessing sync_session which has been destroyed.");
         }
     }
 
-    void sync_session::wait_for_upload_completion(std::function<void(std::error_code)> &&callback) {
+    void sync_session::wait_for_upload_completion(std::function<void(status)> &&callback) {
         if (auto session = m_session.lock()) {
-            session->wait_for_upload_completion([cb = std::move(callback)](auto ec) {
-                cb(ec.get_std_error_code());
+            session->wait_for_upload_completion([cb = std::move(callback)](::realm::Status s) {
+                cb(std::move(s));
             });
         } else {
             throw std::runtime_error("Realm: Error accessing sync_session which has been destroyed.");
@@ -36,12 +39,12 @@ namespace realm::internal::bridge {
         std::future<void> f = p.get_future();
 
         if (auto session = m_session.lock()) {
-            session->wait_for_upload_completion([p = std::move(p)](auto ec) mutable {
-                if (ec.get_std_error_code()) {
-                    p.set_exception(std::make_exception_ptr(ec.get_std_error_code()));
+            session->wait_for_upload_completion([p = std::move(p)](::realm::Status s) mutable {
+                if (s.is_ok()) {
+                    p.set_value();
                 }
                 else {
-                    p.set_value();
+                    p.set_exception(std::make_exception_ptr(s.code()));
                 }
             });
         } else {
@@ -55,12 +58,12 @@ namespace realm::internal::bridge {
         std::promise<void> p;
         std::future<void> f = p.get_future();
         if (auto session = m_session.lock()) {
-            session->wait_for_download_completion([p = std::move(p)](auto ec) mutable {
-                if (ec.get_std_error_code()) {
-                    p.set_exception(std::make_exception_ptr(ec.get_std_error_code()));
+            session->wait_for_download_completion([p = std::move(p)](::realm::Status s) mutable {
+                if (s.is_ok()) {
+                    p.set_value();
                 }
                 else {
-                    p.set_value();
+                    p.set_exception(std::make_exception_ptr(s.code()));
                 }
             });
         } else {